--- client/src/pages/mypage/constants.ts
@@ -1,7 +1,0 @@
-export const SERVICE_ROUTE = {
-  test: 'test',
-  workbook: 'workbook',
-  share: 'share',
-};
-
-export default SERVICE_ROUTE;
